dropdown changes

+++ db/migrate/20240108093911_create_loans.rb
@@ -8,19 +8,19 @@ class CreateLoans < ActiveRecord::Migration[7.0]
       t.date :date_of_birth  # Added: Date of Birth
       t.decimal :amount
       t.string :purpose
-      t.numeric :ssn
-      t.numeric :phone_number
+      t.integer :ssn
+      t.integer :phone_number
       t.string :profession
       t.string :address
-      t.numeric :home_number  # Added: Home Number
+      t.integer :home_number  # Added: Home Number
       t.string :street_address2  # Added: Street Address 2
       t.string :city  # Added: City
       t.string :state  # Added: State
-      t.numeric :zip  # Added: Zip
+      t.integer :zip  # Added: Zip
       t.string :residence_type  # Added: Residence Type
       t.decimal :annual_income  # Added: Annual Income
       t.string :employer_name  # Added: Employer Name
-      t.numeric :employer_phone  # Added: Employer Phone
+      t.integer :employer_phone  # Added: Employer Phone
       
       t.timestamps
     end
